Separate tests by calling contexts.
Currently, we are considering all calling contexts. We now check each one separately. The test parameters now make more sense, but the downside is that now all calls must be consistent. This may cause problems because Python supports union types. I've disabled one test due to this problem, but it's not caused by union types but rather a bug in the tensor dataflow tracking.